+++ activity_dashboard_mngmnt/__manifest__.py
@@ -4,7 +4,7 @@
 #
 #    Cybrosys Technologies Pvt. Ltd.
 #    Copyright (C) 2022-TODAY Cybrosys Technologies (<https://www.cybrosys.com>).
-#    Author: Megha K (<https://www.cybrosys.com>)
+#    Author: Cybrosys Technologies (<https://www.cybrosys.com>)
 #
 #    This program is free software: you can modify
 #    it under the terms of the GNU Affero General Public License (AGPL) as
@@ -20,23 +20,23 @@
 #    along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 ###################################################################################
-
 {
     'name': 'Activity Management',
-    'version': '16.0.1.0.0',
+    'version': '16.0.1.1.0',
     'category': 'Tools',
-    'summary': 'Advance Activity Management and Dashboard View',
+    'summary': 'Advance Activity Management and Dashboard View.',
     'author': 'Cybrosys Techno Solutions',
     'maintainer': 'Cybrosys Techno Solutions',
     'company': 'Cybrosys Techno Solutions',
     'website': 'https://www.cybrosys.com',
-    'description': "Advance Activity Management and Dashboard View",
+    'description': "Advance Activity Management and Dashboard View to track "
+                   "activity of users.",
     'depends': ['base', 'mail'],
     'images': ['static/description/banner.png'],
     'data': [
         'security/ir.model.access.csv',
-        'views/my_activity.xml',
-        'views/activity_tag.xml',
+        'views/mail_activity_views.xml',
+        'views/activity_tag_views.xml',
         'views/activity_dashbord.xml'
     ],
     'assets': {
@@ -49,7 +49,7 @@
         ],
     },
     'installable': True,
-    'application': False,
+    'application': True,
     'auto_install': False,
     'license': 'AGPL-3',
 }