+++ activity_dashboard_mngmnt/models/__init__.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 ###################################################################################
-#    Freight Management
+#    Activity Management
 #
 #    Cybrosys Technologies Pvt. Ltd.
 #    Copyright (C) 2022-TODAY Cybrosys Technologies (<https://www.cybrosys.com>).
-#    Author: Megha K (<https://www.cybrosys.com>)
+#    Author: Cybrosys Technologies (<https://www.cybrosys.com>)
 #
 #    This program is free software: you can modify
 #    it under the terms of the GNU Affero General Public License (AGPL) as
@@ -20,6 +20,5 @@
 #    along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 ###################################################################################
-
+from . import activity_tag
 from . import mail_activity
-from . import activity_type