+++ activity_dashboard_mngmnt/models/activity_tag.py
@@ -4,7 +4,7 @@
 #
 #    Cybrosys Technologies Pvt. Ltd.
 #    Copyright (C) 2022-TODAY Cybrosys Technologies (<https://www.cybrosys.com>).
-#    Author: Megha K (<https://www.cybrosys.com>)
+#    Author: Cybrosys Technologies (<https://www.cybrosys.com>)
 #
 #    This program is free software: you can modify
 #    it under the terms of the GNU Affero General Public License (AGPL) as
@@ -20,21 +20,26 @@
 #    along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 ###################################################################################
-
 from random import randint
 from odoo import fields, models
 
 
 class ActivityTag(models.Model):
+    """Model to add tags to an activity"""
     _name = "activity.tag"
     _description = "Activity Tag"
 
     def _get_default_color(self):
+        """:returns random intrger between 1 and 11"""
         return randint(1, 11)
 
-    name = fields.Char('Tag Name', required=True, translate=True)
-    color = fields.Integer('Color', default=_get_default_color)
+    name = fields.Char(string='Tag Name',
+                       help='Name of the activity tag.',
+                       required=True,
+                       translate=True)
+    color = fields.Integer(string='Color',
+                           help='Field that gives color to tag.',
+                           default=_get_default_color)
 
     _sql_constraints = [
-        ('name_uniq', 'unique (name)', "Tag name already exists !"),
-    ]
+        ('name_uniq', 'unique (name)', "Tag name already exists !"), ]